# docs/openapi.yaml
openapi: 3.0.3
info:
  version: 0.0.1
  title: Task Apis
  description: タスク管理API

servers:
  - url: https://localhost:8080/

paths:
  '/users/{userId}':
    get:
      summary: ユーザー取得
      parameters:
        - $ref: '#/components/parameters/UserIdPathParam'
      responses: 
        '200':
          description: 正常応答
          content: 
            'application/json':
              schema: 
                $ref: '#/components/schemas/User'
  '/users':
    post:
      summary: ユーザー登録
      requestBody: 
        content: 
          'application/json':
            schema: 
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: 正常応答
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/User'
  '/tasks/{taskId}':
    get:
      summary: Task取得
      parameters:
        - $ref: '#/components/parameters/TaskIdPathParam'
      responses:
        '200':
          description: 正常応答
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Task'
  '/tasks':
    post:
      summary: タスク登録
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: 正常応答
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Task'
    patch:
      summary: タスク更新
      parameters:
        - $ref: '#/components/parameters/TaskIdPathParam'
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '200':
          description: 正常応答
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: タスク削除
      parameters:
        - $ref: '#/components/parameters/TaskIdPathParam'
      responses:
        '204':
          description: 正常応答
components:
  schemas: 
    UserId:
      type: string
      format: uuid
    UserName:
      type: string
    TaskId:
      type: string
      format: uuid  
    TaskStatus:
      type: string
      enum:
        - Task
        - IN_PROGRESS
        - DONE
    TaskDescription:
      type: string
    UserCreateRequest:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/UserName'
    User:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UserId'
        name:
          $ref: '#/components/schemas/UserName'
    TaskCreateRequest:
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/UserId'
        description:
          $ref: '#/components/schemas/TaskDescription'
    TaskPatchRequest:
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/UserId'
        status:
          $ref: '#/components/schemas/TaskStatus'
        description:
          $ref: '#/components/schemas/TaskDescription'
    Task:
      type: object
      properties: 
        id:
          $ref: '#/components/schemas/TaskId'
        userId:
          $ref: '#/components/schemas/UserId'
        status:
          $ref: '#/components/schemas/TaskStatus'
        description:
          $ref: '#/components/schemas/TaskDescription'
  parameters: 
    UserIdPathParam:
      name: userId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UserId'
    TaskIdPathParam:
      name: TaskId
      in: path
      schema: 
        $ref: '#/components/schemas/TaskId'
        